# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/account/": {
            "get": {
                "description": "Retrieve current authenticated user data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get user account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update authenticated user account details",
                "consumes": [
                    "application/json",
                    " multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update user account",
                "parameters": [
                    {
                        "description": "Updated account data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_forms.EditAccount"
                        }
                    },
                    {
                        "type": "file",
                        "description": "New avatar image",
                        "name": "avatar",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the authenticated user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Delete user account",
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/recommendation/": {
            "get": {
                "description": "Get personalized event recommendations for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get event recommendations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "User-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backend_internal_models.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_forms.EditAccount": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "birthday": {
                    "description": "1970-01-01",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "backend_internal_models.Address": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "street": {
                    "type": "string"
                },
                "streetNumber": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "eventID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_models.Coords": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.DeviceToken": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "os": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "description": "FMC token",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.User"
                    }
                }
            }
        },
...
        "backend_internal_models.Event": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.Comment"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "eventOrganizers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.User"
                    }
                },
                "event_tags": {
                    "description": "for ml",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.Tag"
                    }
                },
                "event_type": {
                    "$ref": "#/definitions/backend_internal_models.EventType"
                },
                "finishDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "isAdultOnly": {
                    "type": "boolean"
                },
                "isPromoted": {
                    "type": "boolean"
                },
                "location": {
                    "$ref": "#/definitions/backend_internal_models.Location"
                },
                "locationID": {
                    "type": "integer"
                },
                "startDate": {
                    "description": "Timestamp with time zone",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.Vote"
                    }
                }
            }
        },
        "backend_internal_models.EventType": {
            "type": "string",
            "enum": [
                "WORKSHOP",
                "CULTURAL",
                "SPORTS",
                "SOCIAL",
                "COMMUNITY",
                "CHARITY",
                "PARTY"
            ],
            "x-enum-varnames": [
                "Workshop",
                "Cultural",
                "Sports",
                "Social",
                "Community",
                "Charity",
                "Party"
            ]
        },
        "backend_internal_models.Location": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/backend_internal_models.Address"
                },
                "addressID": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "coords": {
                    "$ref": "#/definitions/backend_internal_models.Coords"
                },
                "coordsID": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.ParticipationStatus": {
            "type": "string",
            "enum": [
                "INTERESTED",
                "WILL_PARTICIPATE",
                "NOT_INTERESTED",
                "NOT_APPLICABLE"
            ],
            "x-enum-varnames": [
                "Interested",
                "WillParticipate",
                "NotInterested",
                "NotApplicable"
            ]
        },
...
        "backend_internal_models.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.User": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.Comment"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.DeviceToken"
                    }
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.User"
                    }
                },
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.User"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "isPremium": {
                    "type": "boolean"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/backend_internal_models.Location"
                },
                "locationID": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.VoteAnswer"
                    }
                }
            }
        },
        "backend_internal_models.Vote": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "endDate": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/backend_internal_models.Event"
                },
                "eventID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.VoteOption"
                    }
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "voteType": {
                    "$ref": "#/definitions/backend_internal_models.VoteType"
                }
            }
        },
        "backend_internal_models.VoteAnswer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "optionSelected": {
                    "$ref": "#/definitions/backend_internal_models.VoteOption"
                },
                "optionSelectedId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "voteId": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_models.VoteOption": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "participationStatus": {
                    "$ref": "#/definitions/backend_internal_models.ParticipationStatus"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vote": {
                    "$ref": "#/definitions/backend_internal_models.Vote"
                },
                "voteAnswers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_models.VoteAnswer"
                    }
                },
                "voteId": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_models.VoteType": {
            "type": "string",
            "enum": [
                "CAN_CHANGE_VOTE",
                "CANNOT_CHANGE_VOTE"
            ],
            "x-enum-varnames": [
                "CanChangeVote",
                "CannotChangeVote"
            ]
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}